Fix lint-components by refactoring component code
Fix last lint errors
