+++ comp/process/bundle.go
@@ -18,7 +18,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/process/containercheck/containercheckimpl"
 	"github.com/DataDog/datadog-agent/comp/process/expvars/expvarsimpl"
 	"github.com/DataDog/datadog-agent/comp/process/forwarders/forwardersimpl"
-	"github.com/DataDog/datadog-agent/comp/process/gpusubscriber/gpusubscriberimpl"
+	gpusubscriber "github.com/DataDog/datadog-agent/comp/process/gpusubscriber/fx"
 	"github.com/DataDog/datadog-agent/comp/process/hostinfo/hostinfoimpl"
 	"github.com/DataDog/datadog-agent/comp/process/processcheck/processcheckimpl"
 	"github.com/DataDog/datadog-agent/comp/process/processdiscoverycheck/processdiscoverycheckimpl"
@@ -59,6 +59,6 @@ func Bundle() fxutil.BundleOptions {
 		forwardersimpl.Module(),
 		logscompression.Module(),
 
-		gpusubscriberimpl.Module(),
+		gpusubscriber.Module(),
 	)
 }
