+++ comp/process/gpusubscriber/def/component.go
