+++ comp/process/gpusubscriber/fx-mock/fx.go
@@ -1,0 +1,21 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024-present Datadog, Inc.
+
+// Package fx provides the fx module for the gpu subscriber component
+package fx
+
+import (
+	gpusubscriberimpl "github.com/DataDog/datadog-agent/comp/process/gpusubscriber/impl"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// MockModule defines the fx options for this component.
+func MockModule() fxutil.Module {
+	return fxutil.Component(
+		fxutil.ProvideComponentConstructor(
+			gpusubscriberimpl.NewGpuSubscriberMock,
+		),
+	)
+}
