+++ comp/process/gpusubscriber/impl/noopsubscriber.go
